properties: Move to separate package
Since properties are not part of the OCM wrapper calls, we separate it
into its own package.
