+++ cmd/describe/cluster/cmd.go
@@ -29,8 +29,8 @@ import (
 	clusterprovider "github.com/openshift/rosa/pkg/cluster"
 	"github.com/openshift/rosa/pkg/logging"
 	"github.com/openshift/rosa/pkg/ocm"
-	"github.com/openshift/rosa/pkg/ocm/properties"
 	"github.com/openshift/rosa/pkg/ocm/upgrades"
+	"github.com/openshift/rosa/pkg/properties"
 	rprtr "github.com/openshift/rosa/pkg/reporter"
 )
