+++ pkg/properties/properties.go
